requirements.txt
mcp
duckdb
psycopg2-binary
pytest
pytest-asyncio